complete mypage view
