--- app/src/main/java/com/project/mobo/SignInActivity.kt
@@ -1,23 +1,0 @@
-package com.project.mobo
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import kotlinx.android.synthetic.main.activity_sign_in.*
-
-class SignInActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_sign_in)
-
-        initialUI()
-    }
-
-    private fun initialUI(){
-        btn_sigin_signup.setOnClickListener {
-            val nextIntent = Intent(this, SignUpActivity::class.java)
-            startActivity(nextIntent)
-        }
-    }
-}
