--- app/src/main/java/com/project/mobo/SignUpActivity.kt
@@ -1,12 +1,0 @@
-package com.project.mobo
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class SignUpActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_sign_up)
-    }
-}
