--- app/src/main/java/com/project/mobo/mypage/MyPage_original.kt
@@ -1,79 +1,0 @@
-package com.project.mobo.mypage
-
-import android.Manifest
-import android.annotation.SuppressLint
-import android.app.Activity
-import android.content.Intent
-import android.content.pm.PackageManager
-import android.os.Build.*
-import android.os.Bundle
-import android.widget.Toast
-import androidx.appcompat.app.AppCompatActivity
-import com.project.mobo.R
-import kotlinx.android.synthetic.main.activity_my_page.*
-
-class MyPage_original : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_my_page)
-
-        //Change profile Image
-        profile_image.setOnClickListener {
-            if (VERSION.SDK_INT >= VERSION_CODES.M){
-                if(checkSelfPermission(Manifest.permission.READ_EXTERNAL_STORAGE) ==
-                        PackageManager.PERMISSION_DENIED){
-
-                    val permissions = arrayOf(Manifest.permission.READ_EXTERNAL_STORAGE)
-                    requestPermissions(permissions, PERMISSION_CODE)
-                }
-                else{
-                    pickImageFromGallery()
-                }
-            }
-        }
-    }
-
-    //Change profile image
-    private fun pickImageFromGallery(){
-        val intent = Intent(Intent.ACTION_PICK)
-        intent.type = "image/*"
-        startActivityForResult(intent, IMAGE_PICK_CODE)
-    }
-
-    companion object{
-        private val IMAGE_PICK_CODE = 1000
-        private val PERMISSION_CODE = 1001
-    }
-
-    override fun onRequestPermissionsResult(
-        requestCode: Int,
-        permissions: Array<out String>,
-        grantResults: IntArray
-    ) {
-        when(requestCode){
-            PERMISSION_CODE -> {
-                if (grantResults.size > 0 && grantResults[0] ==
-                        PackageManager.PERMISSION_GRANTED){
-                    pickImageFromGallery()
-                }
-                else{
-                    Toast.makeText(this, "Permission denied", Toast.LENGTH_SHORT).show()
-                }
-            }
-        }
-    }
-
-    @SuppressLint("MissingSuperCall")
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        if(resultCode == Activity.RESULT_OK && requestCode == IMAGE_PICK_CODE){
-            profile_image.setImageURI(data?.data)
-        }
-
-        else if (requestCode == 1){
-            if(resultCode == Activity.RESULT_OK) run {
-                val newName: String? = data?.getStringExtra("data")
-                profile_name.setText(newName)
-            }
-        }
-    }
-}
