+++ app/src/main/res/drawable/edittxt_rectangle.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item>
+        <shape android:shape="rectangle">
+            <solid android:color="#ffffff"/>
+            <stroke
+                android:width="2dp"
+                android:color="#ff6a07"
+                />
+        </shape>
+    </item>
+</selector>
