+++ app/src/main/res/drawable/rectangle_gradation.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <gradient
+        android:angle="45"
+        android:type="linear"
+        android:startColor="#ff6600"
+        android:centerColor="#FF7B23"
+        android:endColor="#ff8f45"/>
+    <corners android:radius="40dp" />
+</shape>
