+++ app/src/main/res/values/strings.xml
@@ -12,6 +12,41 @@
     <string name="interest"># 관심사</string>
     <string name="hobbit">테니스, 독서, 창업</string>
     <string name="advan"># 장점</string>
-    <string name="newPeople">수정</string>
+    <string name="change">수정</string>
     <string name="iden">둥글둥글한 성격</string>
+    <string name="mypage">마이 페이지</string>
+    <string name="btn">버튼</string>
+    <string name="profiletxt">영화 소울메이트를 찾아요 !</string>
+    <string name="dummyname">안형민</string>
+    <string name="dummyoneline">#겁많은 #낯가리는 #성실한 #귀여운 #착한</string>
+    <string name="myticketandpopcorn">나의 보유량</string>
+    <string name="x_2">X 2</string>
+    <string name="todo">TODO</string>
+    <string name="ticket_btn">Ticket Box</string>
+    <string name="dummy_school">고려대학교</string>
+    <string name="dummy_age">26세</string>
+    <string name="dummy_department">미디어학부</string>
+    <string name="dummy_livingarea">서대문구</string>
+    <string name="dummy_name">신가인</string>
+    <string name="dummy_id">아이디</string>
+    <string name="dummy_password">비밀번호</string>
+    <string name="age">나이</string>
+    <string name="school">학교/직장</string>
+    <string name="department">학과</string>
+    <string name="livingarea">생할지역</string>
+    <string name="name">이름</string>
+    <string name="id">아이디</string>
+    <string name="password">비밀번호</string>
+    <string name="textview_information">[개인정보]</string>
+    <string name="matching_gender">매칭 상대 성별</string>
+    <string name="gender_man">남성</string>
+    <string name="gender_woman">여성</string>
+    <string name="gender_no">상관없음</string>
+    <string name="matching_age">매칭 상대 나이대</string>
+    <string name="age_min">최소</string>
+    <string name="age_max">최대</string>
+    <string name="dummy_age_min">20</string>
+    <string name="dummy_age_max">32</string>
+    <string name="end_change">수정 완료</string>
+    <string name="_2">라디오 그룹 버튼2</string>
 </resources>
